% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mapView,whos_sites-method}
\alias{mapView,whos_sites-method}
\title{View \code{whos_sites} object interactively}
\usage{
\S4method{mapView}{whos_sites}(x, ...)
}
\arguments{
\item{x}{An object of class whos_sites.}

\item{...}{Additional arguments passed to \code{mapview::mapView}.}
}
\description{
A method for the generic function \code{mapview::mapView} which
allows users to view WHOS sites interactively.
}
