% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whos_get_values.R
\name{metadata}
\alias{metadata}
\title{Get metadata for object returned by \code{whos_get_values}}
\usage{
metadata(x, ...)
}
\arguments{
\item{x}{Tibble.}

\item{...}{Additional arguments. None implemented.}
}
\value{
list
}
\description{
Get metadata for object returned by \code{whos_get_values}
}
