% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whos_get_values.R
\name{plot.whos_values}
\alias{plot.whos_values}
\title{Plot values}
\usage{
\method{plot}{whos_values}(x, ...)
}
\arguments{
\item{x}{Tibble.}

\item{...}{Additional arguments. None implemented.}
}
\value{
ggplot
}
\description{
Plot values
}
