% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whos-package.R
\docType{package}
\name{whos-package}
\alias{whos-package}
\title{whos: Retrieve data from the WMO Hydrological Observing System}
\description{
Utility functions to download data from the WMO Hydrological Observing System.
}
\examples{
\dontrun{
## Set WHOS ID
my_id <- "whos-8b5fc809-75ef-4877-baff-04ae9e227243"
Sys.setenv(WHOS_ID=my_id)
whos_id()

views <- whos_get_views()
server <- whos_get_server(view = "uk-nrfa")
sites <- whos_get_sites(server)

## To retrieve data, WHOS expects a value from the "SiteCode"
## column. Here we find stations on the River Cherwell:
cherwell_sites <- grep("Cherwell", sites$SiteName)
site_code <- sites[cherwell_sites[1], "SiteCode", drop = TRUE]

## Get list of site_info measured at this site
site_info <- whos_get_site_info(server, site = site_code)
site_info

## Let's retrieve data for rainfall and streamflow
daily_Q_var_code <- site_info$VariableCode[1]
daily_P_var_code <- site_info$VariableCode[8]

streamflow <- whos_get_values(
  server,
  site_code = site_code,
  variable_code = daily_Q_var_code,
  site_info = site_info
)
metadata(streamflow)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonmoulds/whos}
}

}
\author{
\strong{Maintainer}: Simon Moulds \email{sim.moulds@gmail.com} (\href{https://orcid.org/0000-0002-7297-482X}{ORCID})

}
