% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whos_get_site_info.R
\name{whos_get_site_info}
\alias{whos_get_site_info}
\title{Get site information}
\usage{
whos_get_site_info(server, site, ...)
}
\arguments{
\item{server}{Character. WHOS server}

\item{site}{Character. Site code}
}
\value{
tibble
}
\description{
Retrieves metadata for a given WHOS site
}
