% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whos_get_sites.R
\name{whos_get_sites}
\alias{whos_get_sites}
\title{Get sites from WHOS}
\usage{
whos_get_sites(server, ...)
}
\arguments{
\item{server}{The URL of the WHOS server}
}
\value{
A tibble.
}
\description{
This function gets the table of sites from a given WHOS server
}
