% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whos_get_values.R
\name{whos_get_values}
\alias{whos_get_values}
\title{Get values}
\usage{
whos_get_values(server, site_info)
}
\arguments{
\item{server}{Character. WHOS server}

\item{site_info}{Tibble.}
}
\value{
tibble
}
\description{
Retrieve data for a given site.
}
