% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whos_get_server.R
\name{whos_get_server}
\alias{whos_get_server}
\title{Get server}
\usage{
whos_get_server(view)
}
\arguments{
\item{view}{WHOS view.}
}
\value{
Character
}
\description{
Retrieve the URL for a given WHOS view.
}
